Refactor document CLI to use pipeline runner
